requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
discord.py==1.3.3
idna==2.9
lxml==4.5.1
multidict==4.7.6
piazza-api==0.11.0
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
